requirements.txt
python-uinput
xmltodict